+++ src/models/generic_vhost_user_config.rs
@@ -18,9 +18,9 @@ pub struct GenericVhostUserConfig {
     #[serde(rename = "queue_size", skip_serializing_if = "Option::is_none")]
     pub queue_size: Option<Vec<u16>>,
     #[serde(rename = "pci_segment", skip_serializing_if = "Option::is_none")]
-    pub pci_segment: Option<i32>,
+    pub pci_segment: Option<i16>,
     #[serde(rename = "pci_device_id", skip_serializing_if = "Option::is_none")]
-    pub pci_device_id: Option<i32>,
+    pub pci_device_id: Option<u8>,
     #[serde(rename = "virtio_id")]
     pub virtio_id: u32,
 }