+++ src/models/pmem_config.rs
@@ -22,9 +22,9 @@ pub struct PmemConfig {
     #[serde(rename = "discard_writes", skip_serializing_if = "Option::is_none")]
     pub discard_writes: Option<bool>,
     #[serde(rename = "pci_segment", skip_serializing_if = "Option::is_none")]
-    pub pci_segment: Option<i32>,
+    pub pci_segment: Option<i16>,
     #[serde(rename = "pci_device_id", skip_serializing_if = "Option::is_none")]
-    pub pci_device_id: Option<i32>,
+    pub pci_device_id: Option<u8>,
     #[serde(rename = "id", skip_serializing_if = "Option::is_none")]
     pub id: Option<String>,
 }