+++ src/models/vdpa_config.rs
@@ -20,9 +20,9 @@ pub struct VdpaConfig {
     #[serde(rename = "iommu", skip_serializing_if = "Option::is_none")]
     pub iommu: Option<bool>,
     #[serde(rename = "pci_segment", skip_serializing_if = "Option::is_none")]
-    pub pci_segment: Option<i32>,
+    pub pci_segment: Option<i16>,
     #[serde(rename = "pci_device_id", skip_serializing_if = "Option::is_none")]
-    pub pci_device_id: Option<i32>,
+    pub pci_device_id: Option<u8>,
     #[serde(rename = "id", skip_serializing_if = "Option::is_none")]
     pub id: Option<String>,
 }